Readme
This project is a line-following car that uses infrared sensors to detect and follow a black line on a white surface. The car is controlled by an Arduino board and uses the AFMotor library to control the motors. The README file contains information on how to install and use the AFMotor library, as well as instructions on how to use the line-following car. It also includes a circuit diagram and a list of team members who contributed to the project.